feat(envios): crear y editar una zona ocurre en su propia pantalla
El modal se quedaba corto para una escalera de tarifas con rangos: se
cortaba y obligaba a desplazarse dentro de una ventana flotante. Las
zonas pasan a dos rutas propias, con la misma forma que productos,
categorías y combos, y la lista solo lista.
