+++ app/admin/settings/shipping/components/zone-editor.tsx
@@ -1,90 +1,36 @@
 "use client"
 
-import { useEffect, useMemo, useState } from "react"
+import { useState } from "react"
+import Link from "next/link"
 import { useRouter } from "next/navigation"
-import { useFieldArray, useForm, type Control, type FieldPath } from "react-hook-form"
-import { zodResolver } from "@hookform/resolvers/zod"
-import { Check, ChevronsUpDown, Loader2, Pencil, Plus, Save, X } from "lucide-react"
+import { Pencil, Plus, X } from "lucide-react"
 import { toast } from "sonner"
 
 import { ConfirmActionButton } from "@/components/admin/confirm-action-button"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
-import {
-  Command,
-  CommandEmpty,
-  CommandGroup,
-  CommandInput,
-  CommandItem,
-  CommandList,
-} from "@/components/ui/command"
-import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from "@/components/ui/dialog"
-import { FieldError } from "@/components/ui/field-error"
-import { FormField } from "@/components/ui/form-field"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table"
 import { translations } from "@/lib/i18n/translations"
-import type { Department, Municipality } from "@/lib/shipping/locations-api"
 import {
   UNMATCHED_DESTINATION_ACTIONS,
   destinationKey,
-  findShippingLadderGaps,
   shippingRateBasisLabelKey,
-  shippingRateLadderSchema,
-  shippingZoneFormSchema,
-  toRateLadderPayload,
-  type ShippingRateLadder,
   type UnmatchedDestinationAction,
-  type ZoneEditorFormValues,
 } from "@/lib/shipping/schemas"
-import type { MissingWeightProduct, ShippingZoneRecord } from "@/lib/supabase/shipping-zones-api"
-import {
-  deleteShippingZoneAction,
-  listShippingDepartmentsAction,
-  listShippingMunicipalitiesAction,
-  saveShippingZoneAction,
-  updateUnmatchedDestinationActionAction,
-} from "../actions"
-import { EMPTY_RANGE_ROW, RateLadderField } from "./rate-ladder-field"
+import type { ShippingZoneRecord } from "@/lib/supabase/shipping-zones-api"
+import { deleteShippingZoneAction, updateUnmatchedDestinationActionAction } from "../actions"
 
 const copy = translations.es.shipping.zones
 
-// Replaces a plain `string | "new" | null` state: "new" used to be absorbed
-// by `string`, so nothing enforced the distinction and every read site had
-// to re-derive it by hand (was editingZoneId "new", null, or a real id?).
-// A discriminated union makes each of the dialog's three states -- closed,
-// creating, editing a specific zone -- its own checked shape instead.
-type ZoneEditorState = { mode: "closed" } | { mode: "creating" } | { mode: "editing"; zoneId: string }
-
 export function ShippingZonesSection({
   zones,
-  missingWeightProducts,
   unmatchedDestinationAction,
 }: {
   zones: ShippingZoneRecord[]
-  missingWeightProducts: MissingWeightProduct[]
   unmatchedDestinationAction: UnmatchedDestinationAction
 }) {
-  const [editorState, setEditorState] = useState<ZoneEditorState>({ mode: "closed" })
-  const [departments, setDepartments] = useState<Department[]>([])
-
-  useEffect(() => {
-    listShippingDepartmentsAction()
-      .then(setDepartments)
-      .catch((error) => console.error("[Shipping Zones] Error al cargar departamentos:", error))
-  }, [])
-
-  const editingZoneId = editorState.mode === "editing" ? editorState.zoneId : null
-  // Looked up only to PRE-FILL the form -- a stale `zones` list (refreshed
-  // elsewhere while the dialog stayed open) can legitimately miss it, but
-  // that must never flip the save into a create: editingZoneId above, not
-  // this lookup, is what ZoneForm uses to decide create-vs-edit.
-  const editingZone = editingZoneId ? zones.find((zone) => zone.id === editingZoneId) ?? null : null
-
   return (
     <div className="space-y-6">
       <UnmatchedDestinationCard defaultValue={unmatchedDestinationAction} />
@@ -95,51 +41,25 @@ export function ShippingZonesSection({
             <CardTitle className="text-base">{copy.sectionTitle}</CardTitle>
             <CardDescription>{copy.sectionDescription}</CardDescription>
           </div>
-          <Button type="button" size="sm" className="shrink-0 gap-1.5" onClick={() => setEditorState({ mode: "creating" })}>
-            <Plus className="h-4 w-4" /> {copy.addButton}
+          <Button size="sm" className="shrink-0 gap-1.5" asChild>
+            <Link href="/admin/settings/shipping/zones/new">
+              <Plus className="h-4 w-4" /> {copy.addButton}
+            </Link>
           </Button>
         </CardHeader>
         <CardContent>
           {zones.length === 0 ? (
             <p className="py-8 text-center text-sm text-muted-foreground">{copy.emptyDescription}</p>
           ) : (
-            <ZonesTable zones={zones} onEdit={(zoneId) => setEditorState({ mode: "editing", zoneId })} />
+            <ZonesTable zones={zones} />
           )}
         </CardContent>
       </Card>
-
-      <Dialog
-        open={editorState.mode !== "closed"}
-        onOpenChange={(open) => !open && setEditorState({ mode: "closed" })}
-      >
-        <DialogContent className="editor-chrome max-h-[85vh] max-w-2xl overflow-y-auto">
-          <DialogHeader>
-            <DialogTitle>{editorState.mode === "creating" ? copy.createTitle : copy.editTitle}</DialogTitle>
-            <DialogDescription>{copy.sectionDescription}</DialogDescription>
-          </DialogHeader>
-          {editorState.mode !== "closed" ? (
-            <ZoneForm
-              key={editingZoneId ?? "new"}
-              zoneId={editingZoneId}
-              zone={editingZone}
-              departments={departments}
-              missingWeightProducts={missingWeightProducts}
-              onClose={() => setEditorState({ mode: "closed" })}
-            />
-          ) : null}
-        </DialogContent>
-      </Dialog>
     </div>
   )
 }
 
-function ZonesTable({
-  zones,
-  onEdit,
-}: {
-  zones: ShippingZoneRecord[]
-  onEdit: (zoneId: string) => void
-}) {
+function ZonesTable({ zones }: { zones: ShippingZoneRecord[] }) {
   return (
     <Table>
       <TableHeader>
@@ -170,8 +90,10 @@ function ZonesTable({
             </TableCell>
             <TableCell className="text-right">
               <div className="flex items-center justify-end gap-1">
-                <Button variant="outline" size="sm" onClick={() => onEdit(zone.id)} className="gap-1.5">
-                  <Pencil className="h-3.5 w-3.5" /> {copy.editTitle}
+                <Button variant="outline" size="sm" className="gap-1.5" asChild>
+                  <Link href={`/admin/settings/shipping/zones/${zone.id}/edit`}>
+                    <Pencil className="h-3.5 w-3.5" /> {copy.editTitle}
+                  </Link>
                 </Button>
                 <DeleteZoneButton zoneId={zone.id} zoneName={zone.name} />
               </div>
@@ -239,306 +161,5 @@ function UnmatchedDestinationCard({ defaultValue }: { defaultValue: UnmatchedDes
         </Select>
       </CardContent>
     </Card>
-  )
-}
-
-function toRateLadderFormValues(ladder: ShippingRateLadder): ZoneEditorFormValues["rateLadder"] {
-  if (ladder.basis === "flat") {
-    return { basis: "flat", amount: ladder.amount, ranges: [EMPTY_RANGE_ROW] }
-  }
-  return { basis: ladder.basis, amount: "0", ranges: ladder.ranges }
-}
-
-function toDefaultValues(zone: ShippingZoneRecord | null): ZoneEditorFormValues {
-  if (!zone) {
-    return {
-      name: "",
-      destinations: [],
-      rateLadder: { basis: "flat", amount: "0", ranges: [EMPTY_RANGE_ROW] },
-    }
-  }
-
-  return {
-    name: zone.name,
-    destinations: zone.destinations.map((destination) => ({
-      departmentCode: destination.departmentCode,
-      municipalityCode: destination.municipalityCode,
-    })),
-    rateLadder: toRateLadderFormValues(zone.rateLadder),
-  }
-}
-
-function nameHintKey(kind: "department" | "municipality", code: string): string {
-  return `${kind}:${code}`
-}
-
-function buildNameHints(zone: ShippingZoneRecord | null): Map<string, string> {
-  const hints = new Map<string, string>()
-  for (const destination of zone?.destinations ?? []) {
-    hints.set(nameHintKey("department", destination.departmentCode), destination.departmentName)
-    if (destination.municipalityCode && destination.municipalityName) {
-      hints.set(nameHintKey("municipality", destination.municipalityCode), destination.municipalityName)
-    }
-  }
-  return hints
-}
-
-function ZoneForm({
-  zoneId,
-  zone,
-  departments,
-  missingWeightProducts,
-  onClose,
-}: {
-  // Decides create-vs-edit on submit -- sourced from editorState, never
-  // re-derived from `zone` (see the editingZoneId comment above the caller).
-  zoneId: string | null
-  zone: ShippingZoneRecord | null
-  departments: Department[]
-  missingWeightProducts: MissingWeightProduct[]
-  onClose: () => void
-}) {
-  const router = useRouter()
-  const nameHints = useMemo(() => buildNameHints(zone), [zone])
-
-  const {
-    control,
-    register,
-    handleSubmit,
-    setError,
-    formState: { errors, isSubmitting },
-  } = useForm<ZoneEditorFormValues>({
-    resolver: zodResolver(shippingZoneFormSchema),
-    defaultValues: toDefaultValues(zone),
-  })
-
-  // Takes handleSubmit's own validated payload instead of re-reading the
-  // form with getValues(): one reading of the form, not two that can quietly
-  // differ (the resolver trims `name`; a second, untrimmed read would have
-  // undone that).
-  async function onSubmit(values: ZoneEditorFormValues) {
-    const ladderPayload = toRateLadderPayload(values.rateLadder)
-    const parsedLadder = shippingRateLadderSchema.safeParse(ladderPayload)
-    if (!parsedLadder.success) {
-      // looseRateLadderFieldSchema (the resolver's own schema) never fails on
-      // the ladder, so this is the only place a broken row/amount is ever
-      // caught -- surfaced on the exact field RateLadderField renders
-      // (rate-ladder-field.tsx's Controllers read fieldState off this SAME
-      // control), the same "inline error, no toast" treatment name/destinations
-      // already get from the resolver above.
-      for (const issue of parsedLadder.error.issues) {
-        setError(`rateLadder.${issue.path.join(".")}` as FieldPath<ZoneEditorFormValues>, {
-          type: "manual",
-          message: issue.message,
-        })
-      }
-      return
-    }
-
-    const gapIssues = findShippingLadderGaps(parsedLadder.data)
-    if (gapIssues.length > 0) {
-      // D6 is a cross-row invariant, not one field's fault -- one toast per
-      // issue (same "several problems, several toasts" shape processOrder
-      // already uses in app/checkout/page.tsx for stock errors) so fixing the
-      // first one doesn't just uncover a second the owner was never shown.
-      gapIssues.forEach((issue) => toast.error(issue.message))
-      return
-    }
-
-    const result = await saveShippingZoneAction(
-      { name: values.name, destinations: values.destinations, rateLadder: parsedLadder.data },
-      zoneId ?? undefined,
-    )
-
-    if (!result.success) {
-      toast.error(result.error || copy.saveErrorToast)
-      return
-    }
-
-    toast.success(copy.savedToast)
-    router.refresh()
-    onClose()
-  }
-
-  return (
-    <form onSubmit={(event) => void handleSubmit(onSubmit)(event)} className="space-y-4">
-      <FormField id="zone-name" label={copy.nameLabel} error={errors.name?.message}>
-        {(field) => <Input {...field} placeholder={copy.namePlaceholder} {...register("name")} />}
-      </FormField>
-
-      <DestinationsField control={control} departments={departments} nameHints={nameHints} />
-      <FieldError message={errors.destinations?.message} />
-
-      <RateLadderField control={control} missingWeightProducts={missingWeightProducts} />
-
-      <div className="flex flex-col-reverse gap-2 sm:flex-row sm:justify-end">
-        <Button type="button" variant="outline" onClick={onClose} disabled={isSubmitting}>
-          {copy.cancelButton}
-        </Button>
-        <Button type="submit" disabled={isSubmitting} className="gap-2">
-          {isSubmitting ? <Loader2 className="h-4 w-4 animate-spin" /> : <Save className="h-4 w-4" />}
-          {isSubmitting ? copy.saving : copy.saveButton}
-        </Button>
-      </div>
-    </form>
-  )
-}
-
-function DestinationsField({
-  control,
-  departments,
-  nameHints,
-}: {
-  control: Control<ZoneEditorFormValues>
-  departments: Department[]
-  nameHints: Map<string, string>
-}) {
-  const { fields, append, remove } = useFieldArray({ control, name: "destinations" })
-  const [names, setNames] = useState(nameHints)
-  const [selectedDepartment, setSelectedDepartment] = useState("")
-  const [municipalities, setMunicipalities] = useState<Municipality[]>([])
-  const [loadingMunicipalities, setLoadingMunicipalities] = useState(false)
-  const [municipalityPickerOpen, setMunicipalityPickerOpen] = useState(false)
-
-  // Fetches, not a state mirror: the effect's only job is asking the server
-  // for the chosen department's municipios, so it stays an effect. Every
-  // setState the effect body itself would need synchronously (clearing the
-  // previous department's stale list, flagging the fetch as loading) instead
-  // happens in the SELECT handler below that causes it -- only the async
-  // .then/.finally callbacks reporting the fetch's own outcome set state
-  // from here, the "derive it from the event that caused it" shape React's
-  // own effect guidance asks for.
-  useEffect(() => {
-    if (!selectedDepartment) return
-
-    let active = true
-    listShippingMunicipalitiesAction(selectedDepartment)
-      .then((result) => {
-        if (active) setMunicipalities(result)
-      })
-      .catch((error) => console.error("[Shipping Zones] Error al cargar municipios:", error))
-      .finally(() => {
-        if (active) setLoadingMunicipalities(false)
-      })
-
-    return () => {
-      active = false
-    }
-  }, [selectedDepartment])
-
-  function selectDepartment(departmentCode: string) {
-    setSelectedDepartment(departmentCode)
-    setMunicipalities([])
-    setLoadingMunicipalities(true)
-  }
-
-  // departmentName falls back to the `departments` prop directly instead of
-  // seeding it into `names` on every prop change: `names` only needs to hold
-  // hints the prop can't supply (a destination's name once picked, or an
-  // already-saved zone's own destinations).
-  function departmentName(departmentCode: string): string {
-    return (
-      names.get(nameHintKey("department", departmentCode)) ??
-      departments.find((department) => department.code === departmentCode)?.name ??
-      departmentCode
-    )
-  }
-
-  const existingKeys = new Set(fields.map(destinationKey))
-
-  function addWholeDepartment() {
-    if (!selectedDepartment || existingKeys.has(destinationKey({ departmentCode: selectedDepartment, municipalityCode: null }))) return
-    append({ departmentCode: selectedDepartment, municipalityCode: null })
-  }
-
-  function addMunicipality(municipality: Municipality) {
-    if (existingKeys.has(destinationKey({ departmentCode: municipality.departmentCode, municipalityCode: municipality.code }))) return
-
-    setNames((current) => {
-      const next = new Map(current)
-      next.set(nameHintKey("department", municipality.departmentCode), municipality.departmentName)
-      next.set(nameHintKey("municipality", municipality.code), municipality.name)
-      return next
-    })
-    append({ departmentCode: municipality.departmentCode, municipalityCode: municipality.code })
-    setMunicipalityPickerOpen(false)
-  }
-
-  return (
-    <div className="space-y-3">
-      <Label>{copy.destinationsLabel}</Label>
-
-      <div className="flex flex-col gap-2 sm:flex-row">
-        <Select value={selectedDepartment} onValueChange={selectDepartment}>
-          <SelectTrigger className="w-full sm:w-64">
-            <SelectValue placeholder={copy.departmentPlaceholder} />
-          </SelectTrigger>
-          <SelectContent className="editor-chrome">
-            {departments.map((department) => (
-              <SelectItem key={department.code} value={department.code}>
-                {department.name}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-        <Button type="button" variant="outline" size="sm" disabled={!selectedDepartment} onClick={addWholeDepartment}>
-          {copy.addWholeDepartmentButton}
-        </Button>
-      </div>
-
-      {selectedDepartment ? (
-        <Popover open={municipalityPickerOpen} onOpenChange={setMunicipalityPickerOpen}>
-          <PopoverTrigger asChild>
-            <Button type="button" variant="outline" size="sm" className="w-full justify-between sm:w-72">
-              {copy.addMunicipalityButton}
-              <ChevronsUpDown className="h-4 w-4 shrink-0 opacity-50" />
-            </Button>
-          </PopoverTrigger>
-          <PopoverContent className="editor-chrome w-72 p-0">
-            <Command>
-              <CommandInput placeholder={copy.municipalitySearchPlaceholder} />
-              <CommandList>
-                <CommandEmpty>{loadingMunicipalities ? copy.loadingMunicipalities : copy.noMunicipalitiesFound}</CommandEmpty>
-                <CommandGroup>
-                  {municipalities.map((municipality) => {
-                    const alreadyAdded = existingKeys.has(
-                      destinationKey({ departmentCode: municipality.departmentCode, municipalityCode: municipality.code }),
-                    )
-                    return (
-                      <CommandItem
-                        key={municipality.id}
-                        value={municipality.name}
-                        disabled={alreadyAdded}
-                        onSelect={() => addMunicipality(municipality)}
-                      >
-                        <Check className={alreadyAdded ? "mr-2 h-4 w-4 opacity-100" : "mr-2 h-4 w-4 opacity-0"} />
-                        {municipality.name}
-                      </CommandItem>
-                    )
-                  })}
-                </CommandGroup>
-              </CommandList>
-            </Command>
-          </PopoverContent>
-        </Popover>
-      ) : null}
-
-      <div className="flex flex-wrap gap-2">
-        {fields.map((field, index) => {
-          const label = field.municipalityCode
-            ? names.get(nameHintKey("municipality", field.municipalityCode)) ?? field.municipalityCode
-            : `${copy.wholeDepartmentPrefix} ${departmentName(field.departmentCode)}`
-
-          return (
-            <Badge key={field.id} variant="secondary" className="gap-1.5 normal-case">
-              {label}
-              <button type="button" onClick={() => remove(index)} aria-label={copy.removeDestinationLabel}>
-                <X className="h-3 w-3" />
-              </button>
-            </Badge>
-          )
-        })}
-      </div>
-    </div>
   )
 }
